% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tab.R
\name{check_tab}
\alias{check_tab}
\title{Check a matrix or data frame}
\usage{
check_tab(tab, num.cols)
}
\arguments{
\item{tab}{Matrix-like object with statistical and p-value columns. Only the signs of the statistics columns are used.
\code{tab} should have non-duplicated row names and should not have missing values.}

\item{num.cols}{Vector of column indices or names that should be numeric.}
}
\description{
Check that an object is a matrix or data frame and ensure it contains the specified columns and that they are numeric.
If so, it returns \code{TRUE}. If not, it triggers an error.
}
