% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_glm.R
\name{mediate_glm}
\alias{mediate_glm}
\title{Mediation analysis using ScreenDMT with glm}
\usage{
mediate_glm(
  E,
  M,
  Y,
  covariates = NULL,
  fam = "gaussian",
  reorder.rows = TRUE,
  p.adj.rate = c("FDR", "FWER"),
  verbose = TRUE,
  check.names = TRUE
)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample of mediators.
Must have more than one feature and have row names and column names.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes. Only continuous, normally distributed outcomes
currently supported.}

\item{covariates}{Numeric vector with one element per sample or matrix-like object with rows corresponding
to samples and columns to covariates to be adjusted for.}

\item{fam}{Character string of family to use in generalized linear model of \code{Y}. The default
\code{"gaussian"} reduces to the usual linear regression model. See stats::family.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{p.adj.rate}{Either "FDR" for false discovery rate or "FWER" for family-wise error rate, the rate controlled by the Bonferroni procedure.}

\item{verbose}{Logical; should messages be given for lack of association between \code{E} & \code{Y} and filtering?}

\item{check.names}{Logical; should \code{names(E)==colnames(M) & colnames(M)==names(Y)} be checked?}
}
\value{
Data frame with columns
\describe{
\item{EMY.chisq}{Overall chi-square for mediation on 1 degreee of freedom.}
\item{EMY.p}{Overall p-value for mediation}
\item{EMY.FDR}{Overall FDR for mediation}
\item{EM.z}{z-score for E-->M, not accounting for direction}
\item{EM.p}{p-value for E-->M, not accounting for direction}
\item{MY.z}{z-score for M-->Y, not accounting for direction}
\item{MY.p}{p-value for M-->Y, not accounting for direction}
}
}
\description{
Mediation analysis for many analytes with ScreenDMT. It tests if rows of \code{M} mediate the effect of exposure \code{E} on outcome
\code{Y}. The associations are tested using base R's \code{glm} or \code{lm}.
Before applying, you should know a priori the direction that \code{E} changes \code{Y} and verify
that \code{E} changes \code{Y} significantly and in the same direction here.
}
\details{
\code{E} and \code{Y} cannot have \code{NA}s. \code{M} may have some \code{NA}s.

Larger chi-square values are more significant.
}
