% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt.R
\name{dmt}
\alias{dmt}
\title{Directional MaxP Test}
\usage{
dmt(
  tab,
  cols = 1:4,
  prod.sgn = 1,
  reorder.rows = FALSE,
  fdr.method = c("BH", "BY"),
  prefix = NULL,
  keep.input = FALSE
)
}
\arguments{
\item{tab}{Matrix-like object with statistical and p-value columns. Only the signs of the statistics columns are used.
\code{tab} should have non-duplicated row names and should not have missing values.}

\item{cols}{Vector of column indices or names in the order of \code{c(stat1, p1, stat2, p2)}.}

\item{prod.sgn}{Either \code{1} or \code{-1}. The sign of the product of the statistics under the alternative hypothesis. For replication, this should be \code{1}.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{fdr.method}{Character string; either "BH" for Benjamini-Hochberg or "BY" for Benjamini-Yekutieli.
See stats::p.adjust.}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{keep.input}{Logical, should returned data frame include \code{tab[, cols]}? If so, \code{tab} must have column names and these
cannot include "chisq", "p", or "FDR".}
}
\value{
Data frame whose rows correspond to the rows of \code{tab} with the same row names and whose columns include
\describe{
\item{chisq}{Chi-square on 1 degreee of freedom.}
\item{p}{P-value}
\item{FDR}{Adjusted p-value}
}
}
\description{
Directional MaxP Test for two studies with signed statistics and p-values supplied via \code{tab}. This can be used to find rows where
there is replication in a common direction or where there is mediation.
}
\details{
Larger chi-square values are more significant.
}
